requirements.txt
# ---------- FastAPI Backend ----------
fastapi
uvicorn[standard]

# Database
sqlalchemy>=2.0
alembic
psycopg2-binary    # only used if connecting to Supabase (safe to keep)
python-dotenv

# Auth (minimal)
python-jose
passlib[bcrypt]

# HTTP requests
requests

# Video Downloading
yt-dlp

# Google Gemini
google-generativeai

# ---------- Streamlit Frontend ----------
streamlit

# ---------- Utilities ----------
pydantic